requirements.txt
flask==2.2.3
GitPython==3.1.31
Werkzeug==2.2.3
Jinja2==3.1.2
requests==2.28.2
pygments==2.15.0
astroid==2.15.0
pylint==2.17.0
markdown==3.4.3 